Solution Provided

+++ Final Project/Frontend/src/app/Restaurants/display-restaurant/display-restaurant.component.ts
@@ -3,12 +3,13 @@ import { FormArray, FormBuilder, FormControl, FormGroup, SelectMultipleControlVa
 import { MatDialog } from '@angular/material/dialog';
 import { MatSelectionList } from '@angular/material/list';
 import { DomSanitizer } from '@angular/platform-browser';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Menu } from 'src/app/Model-Classes/Menu';
 import { Order } from 'src/app/Model-Classes/Order';
-import { Restaurant } from 'src/app/Model-Classes/Restaurant';
 import { OrderComponent } from 'src/app/order/order.component';
+import { Restaurant } from 'src/app/Model-Classes/Restaurant';
 import { LinkService } from 'src/app/Services/link.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-display-restaurant',
@@ -16,11 +17,10 @@ import { LinkService } from 'src/app/Services/link.service';
   styleUrls: ['./display-restaurant.component.css']
 })
 export class DisplayRestaurantComponent implements OnInit {
-
-  RestForm:FormGroup;
-  thumbnail: any;
+RestForm:FormGroup;
+thumbnail: any;
   constructor(private formbuilder:FormBuilder,private link:LinkService,private activate:ActivatedRoute,private sanitizer:DomSanitizer,
-    public dialog: MatDialog) { 
+    public dialog: MatDialog,private router:Router) { 
     this.RestForm = this.formbuilder.group({
       restId: new FormControl('',[Validators.required]),
       restName: new FormControl('',[Validators.required]),
@@ -80,36 +80,51 @@ myData:any;myData1:any;
      this.cartData=count;
    }
   }
+  total:number=0;
   getCount(itemName) {
+    
     return this.menuList1.filter(o => o.itemName === itemName).length;
   }
   menuList1:Menu[]=[]
   increase(pass:any,event){
-    console.log("data",event);
+    
   
       this.menuList1.push(pass);
+      this.price();
       console.log("Menulist",this.menuList1);
   }
-total:number=0;
+  totalCost:any;
+price(){
+  if(this.menuList1.length==0){
+    this.totalCost=0;
+  }
+  this.totalCost=this.menuList1.map(x=> x.price).reduce((x,y)=> x+y);
+}
 
 decrease(pass:any,index:number){
   
     // this.menuList1.pop();
     if(index!==-1){
-      this.menuList1.splice(index,1)
+      this.menuList1.splice(index,1);
+      this.price();
     console.log("Menulist Delete",this.menuList1);
     }
     
 }
 order:Order=new Order();
 orderPlace(){
-  // this.link.placeOrder(this.menuList1).subscribe((x)=>{
-  //   console.log("Order Details",x);
-   
-   // this.order=x;
-    localStorage.setItem("list",JSON.stringify(this.menuList1))
-    
-//  })
+
+  Swal.fire({
+    title: 'After Placing Order You Will Not able to make Changes... Please Confirm...!!!',
+    showDenyButton: true,
+    showCancelButton: true,
+    confirmButtonText: 'Place Order???',
+    denyButtonText: `No`,
+  }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+    if (result.isConfirmed) {
+      Swal.fire('Proceeding to Payment...!', '', 'info')
+      localStorage.setItem("list",JSON.stringify(this.menuList1))
 var l = localStorage.getItem("list");
 console.log("myList1",l);
 
@@ -121,6 +136,15 @@ console.log("myList1",l);
   dialogRef.afterClosed().subscribe(result => {
     console.log('The Dialog was closed!');
   })
+
+
+
+    } else if (result.isDenied) {
+      Swal.fire('Please Confirm Your Menu Choice...!!!', '', 'info')
+    }
+  })
+
+    
 }
 addToFav(){
   // this.RestForm.get('restId')?.setValue(pass.restId);
@@ -136,7 +160,6 @@ addToFav(){
  })
 }
 logout(){
-  
+  this.router.navigateByUrl('/restaurants-details')
 }
-
 }
